migrating comments and comment list to a generic space for use in mug comm and event testing
